Add files via upload
Here's the updated version of the previous file I uploaded -- Final Exercise, with some modifications, in particular selecting only the variables of interest and displaying the plots as far as the linearity assumption is concerned. I also deleted the checking of homoscedasticity since the plots are clearly non-normal and nonlinear and therefore violate the assumptions under parametric. Hence, I decided to directly compute the Kendall's tau correlation coefficient for each paired variables.